remove getters from Builder
