+++ src/main/java/example/lsp/initialPlans/ExampleSchedulingOfTransportChain.java
@@ -248,7 +248,7 @@ import lsp.resources.LSPResource;
 		Random rand = new Random(1);
 		 for(int i = 1; i < 6; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = rand.nextInt(10);
 	        	builder.setCapacityDemand(capacityDemand);
