+++ src/main/java/example/lspAndDemand/requirementsChecking/ExampleCheckRequirementsOfAssigner.java
@@ -3,7 +3,7 @@ package example.lspAndDemand.requirementsChecking;
 import lsp.*;
 import lsp.resources.LSPResource;
 import lsp.shipment.LSPShipment;
-import lsp.shipment.ShipmentUtils;
+import lsp.shipment.LSPShipmentImpl;
 import lsp.usecase.UsecaseUtils;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
@@ -133,7 +133,7 @@ class ExampleCheckRequirementsOfAssigner {
 	    
 	    for(int i = 1; i < 11; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
         	int capacityDemand = rand.nextInt(10);
         	builder.setCapacityDemand(capacityDemand);
