+++ src/main/java/lsp/usecase/LSPShipmentMaker.java
@@ -1,7 +1,7 @@
 package lsp.usecase;
 
 import lsp.shipment.LSPShipment;
-import lsp.shipment.ShipmentUtils;
+import lsp.shipment.LSPShipmentImpl;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
@@ -25,7 +25,7 @@ class LSPShipmentMaker {
 		ArrayList<LSPShipment> shipments = new ArrayList<>();
 
 		for(int i = 0; i < 8; i++){
-			ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(Id.create("Shipment " + i, LSPShipment.class ) );
+			LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(Id.create("Shipment " + i, LSPShipment.class ) );
 			builder.setDeliveryServiceTime(180 );
 			builder.setCapacityDemand(1);
 			TimeWindow startTimeWindow = TimeWindow.newInstance(0, Double.MAX_VALUE);
