+++ src/test/java/lsp/usecase/CollectionLSPSchedulingTest.java
@@ -100,7 +100,7 @@ public class CollectionLSPSchedulingTest {
 
 		for(int i = 1; i < 2; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
         	Random random = new Random(1);
         	int capacityDemand = random.nextInt(4);
         	builder.setCapacityDemand(capacityDemand);
