+++ src/test/java/lsp/usecase/MultipleShipmentsCollectionLSPSchedulingTest.java
@@ -101,7 +101,7 @@ public class MultipleShipmentsCollectionLSPSchedulingTest {
 
 		for(int i = 1; i < 100; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
         	Random random = new Random(1);
         	int capacityDemand = random.nextInt(4);
         	builder.setCapacityDemand(capacityDemand);
