+++ src/test/java/lsp/usecase/SecondReloadLSPSchedulingTest.java
@@ -195,7 +195,7 @@ public class SecondReloadLSPSchedulingTest {
 		
 		 for(int i = 1; i < 2; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = new Random().nextInt(10);
 	        	builder.setCapacityDemand(capacityDemand);
