+++ src/test/java/lspMobsimTests/MultipleIterationsMainRunLSPMobsimTest.java
@@ -167,7 +167,7 @@ public class MultipleIterationsMainRunLSPMobsimTest {
 		int numberOfShipments = 1 + new Random().nextInt(50);
 		for(int i = 1; i < 1 + numberOfShipments; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = 1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
