+++ src/test/java/lspMobsimTests/MultipleItreationsCompleteLSPMobsimTest.java
@@ -229,7 +229,7 @@ public class MultipleItreationsCompleteLSPMobsimTest {
 		
 		for(int i = 1; i < 1 +  numberOfShipments; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand =  1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
