+++ src/test/java/lspScoringTests/MultipleIterationsCollectionLSPScoringTest.java
@@ -10,7 +10,7 @@ import lsp.*;
 import lsp.functions.*;
 import lsp.replanning.LSPReplanningUtils;
 import lsp.scoring.LSPScoringUtils;
-import lsp.shipment.ShipmentUtils;
+import lsp.shipment.LSPShipmentImpl;
 import lsp.usecase.*;
 import org.junit.Before;
 import org.junit.Test;
@@ -123,7 +123,7 @@ public class MultipleIterationsCollectionLSPScoringTest {
 
 		for (int i = 1; i < (numberOfShipments + 1); i++) {
 			Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-			ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+			LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
 			Random random = new Random(1);
 			int capacityDemand = random.nextInt(10);
 			builder.setCapacityDemand(capacityDemand);
