+++ src/test/java/testLSPWithCostTracker/CollectionTrackerTest.java
@@ -2,6 +2,7 @@ package testLSPWithCostTracker;
 
 import lsp.*;
 import lsp.controler.LSPModule;
+import lsp.shipment.LSPShipmentImpl;
 import org.matsim.contrib.freight.events.eventsCreator.LSPEventCreatorUtils;
 import lsp.functions.LSPInfo;
 import lsp.functions.LSPInfoFunctionValue;
@@ -9,7 +10,6 @@ import lsp.replanning.LSPReplanningUtils;
 import lsp.resources.LSPResource;
 import lsp.scoring.LSPScoringUtils;
 import lsp.shipment.LSPShipment;
-import lsp.shipment.ShipmentUtils;
 import lsp.controler.LSPSimulationTracker;
 import lsp.usecase.*;
 import org.junit.Before;
@@ -128,7 +128,7 @@ public class CollectionTrackerTest {
 
 		for(int i = 1; i < 2; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
+        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
         	Random random = new Random(1);
         	int capacityDemand = random.nextInt(4);
         	builder.setCapacityDemand(capacityDemand);
